Update for Scala 2.12 & bump version 0.4.
This is a minimal update to build sumac w/ 2.12.  The core module
compiles and passes tests.  I did not migrate the zk module (code still
there, just commented out of build).

This also updated the version of sbt & scalatest.

I have not made any attempt to cross-compile here.

I chose to bump the version to 0.4-SNAPSHOT, as I figure we could
continue to maintain < 2.12 on the 0.3 line, if there is a need.