Scala 2.12 & bump version to 0.4 (#58)
* Scala 2.12 & bump version to 0.4

This is a minimal update to build sumac w/ 2.12. The core & ext modules
compiles and passes tests. I did not migrate the zk module (code still
there, just commented out of build).  Something seems to have changed
in the apis lowest version of util-zk for 2.12.

I deleted ArgApp, it was deprecated before anyway (you should use
ArgMain).

There was a strange test failure in the SumacSupportTest (scalatra +
swagger).  I made a change to get the test to pass, but I am not really
sure if its right.  I left a TODO, I figure that can be fixed later.

Cross-compilation doesn't work (I dont' know why, `+compile` tries to
compile against scala 2.10).  But manually setting scala version to
2.11.0 or 2.12.8 both work (compiles & tests pass).

I chose to bump the version to 0.4-SNAPSHOT, as I figure we could
continue to maintain < 2.10 on the 0.3 line, if there is a need.

* bump some versions (brings back dependencyTree)

* remove ext-zk module completely

* fix swagger tests properly

* cross compile working